### 2022-02-24 #4201 @editorialbot
[PRE REVIEW]: RIAssigner: A package for gas chromatographic retention index calculation
**Submitting author:** @hechth (<a href="http://orcid.org/0000-0001-6744-996X">Helge Hecht</a>)
**Repository:** https://github.com/RECETOX/RIAssigner
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.3.2
