### 2022-02-24 #4201 @hechth
As reviewers I'd like to suggest yguitton, bovee or jspaezp - they are on the list of registered reviewers for JOSS.

Other editor suggestions would be Mark Jensen, Kristina Riemer or Charlotte Soneson.
