### 2022-02-27 #4201 @arfon
@hechth – as this is on the smaller side of what we allow in JOSS, I'm going to ask the wider editorial team to discuss if it meets the [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) criterion for review by JOSS. We should get back to you sometime next couple of weeks.
