### 2022-03-11 #4201 @Kevin-Mattheus-Moerman
@hechth I am the AEiC on call for JOSS this week. The JOSS board has determined this submission may be in scope for JOSS. I have therefore removed the query scope label, and we'll therefore proceed to find and assign a handling editor.
