### 2022-04-12 #4201 @hechth
@dhhagan Other possible reviewers not in the list of JOSS are:
Theodore Alexandrov (theodev)
Andrew Palmer (andy-d-palmer)
Simon Rogers (sdrogers)
Ralf J. Weber (RJMW)

Other JOSS reviewers which could fit are (github handles):
askerdb
khinsen
davidegraff
dominiquesydow
darinddv
cthoyt
