### 2022-04-22 #4201 @dhhagan
@davidegraff I would start by taking a look at the docs [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html) and [review guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html). I'm also happy to help throughout the process if you have any questions (just ping me here in a comment!).
