### 2022-04-22 #4337 @editorialbot
[REVIEW]: RIAssigner: A package for gas chromatographic retention index calculation
**Submitting author:** @hechth (<a href="http://orcid.org/0000-0001-6744-996X">Helge Hecht</a>)
**Repository:** https://github.com/RECETOX/RIAssigner
**Branch with paper.md** (empty if default branch): paper
**Version:** 0.3.3
