### 2022-05-23 RECETOX/RIAssigner#88 @davidegraff
[JOSS Review] Functionality documentation: linting/style
*cc https://github.com/openjournals/joss-reviews/issues/4337*

The code is documented and largely formatted with `flake8` but there are a few nitpicks:
- ```$ flake8 --ignore E501 RIAssigner > flake8.log```  still outputs some minor errors:
([flake8.log](https://github.com/RECETOX/RIAssigner/files/8755253/flake8.log))
- `assert`s are used when a `ValueError` should be used instead (e.g., [1](https://github.com/RECETOX/RIAssigner/blob/main/RIAssigner/compute/ComputationMethod.py#L29), [2](https://github.com/RECETOX/RIAssigner/blob/main/RIAssigner/data/SimpleData.py#L29), and [3](https://github.com/RECETOX/RIAssigner/blob/main/RIAssigner/data/PandasData.py#L36). `assert`s should be avoided for client-side error-checking (i.e., invalid input values) because [they can be entirely disabled](https://stackoverflow.com/questions/1273211/disable-assertions-in-python)
