### 2022-05-23 RECETOX/RIAssigner#89 @davidegraff
[JOSS Review] Automated tests: some automated tests fail
*cc https://github.com/openjournals/joss-reviews/issues/4337*

After setting up the `riassigner-dev` environment and installing `RIAssigner` according to the [instructions in the README](https://github.com/RECETOX/RIAssigner/#installation), I ran:

```$ pytest > pytest.log```

and some of the tests failed. I have attached the output here: [pytest.log](https://github.com/RECETOX/RIAssigner/files/8755334/pytest.log)
