### 2022-05-24 RECETOX/RIAssigner#88 @hechth
@davidegraff thank you very much for the useful hint - I wasn't aware of that regarding asserts - will fix that accordingly!
