+++ conda/environment.yml
@@ -5,8 +5,8 @@ channels:
   - defaults
 dependencies:
   - numpy
-  - matchms >=0.9.1
+  - matchms >=0.14.0
   - pandas
   - pint >= 0.17
-  - python >=3.7,<3.9
+  - python >=3.7
   - scipy
