### 2022-05-26 RECETOX/RIAssigner#89 @hechth
@davidegraff please check out the PR #90 - this should solve those issues.

Please feel free to comment here or on the PR! Thank you very much!
