### 2022-05-26 7a8c73e
convert non-test asserts to exceptions
