+++ RIAssigner/data/PandasData.py
@@ -33,7 +33,8 @@ class PandasData(Data):
 
     def write(self, filename: str):
         """ Write data on disk. Currently supports 'csv' and 'tsv' formats. """
-        assert filename.endswith((".csv", ".tsv")), "File extension must be 'csv' or 'tsv'."
+        if not filename.endswith((".csv", ".tsv")):
+            raise ValueError("File extension must be 'csv' or 'tsv'.")
         separator = define_separator(filename)
         self._data.to_csv(filename, index=False, sep=separator)
 
