+++ RIAssigner/data/SimpleData.py
@@ -17,17 +17,20 @@ class SimpleData(Data):
             retention_times (Iterable[float]): Retention time values
         """
         super().__init__(None, None, rt_unit)
-        self._assert_valid_input(retention_times, retention_indices)
+        self._validate_input(retention_times, retention_indices)
 
         self._read(retention_times, retention_indices)
 
-    def _assert_valid_input(self, retention_times, retention_indices):
-        assert all(map(Data.is_valid, retention_times)), "Invalid retention time data."
-        assert is_sorted(retention_times), "Retention time data has to be sorted."
-
+    def _validate_input(self, retention_times, retention_indices):
+        if not all(map(Data.is_valid, retention_times)):
+            raise ValueError("Retention time data is invalid.")
+        if not is_sorted(retention_times):
+            raise ValueError("Retention time data has to be sorted.")
         if retention_indices is not None:
-            assert len(retention_times) == len(retention_indices), "Retention time and index data are of different length."
-            assert is_sorted(retention_indices), "Retention index data has to be sorted."
+            if len(retention_times) != len(retention_indices):
+                raise ValueError("Retention times and index data are of different length.")
+            if not is_sorted(retention_indices):
+                raise ValueError("Retention indices data has to be sorted.")
 
     def _read(self, retention_times, retention_indices):
         self._retention_times = Data.URegistry.Quantity(retention_times, self._unit)
