### 2022-05-31 9062d52
rev and minimize codeql config
