### 2022-06-29 #4337 @hechth
@dhhagan I just wanted to ask if there are some updates? The initial submission of the paper is now over 4 months ago with the review being in the 11th week (guidelines state that the reviews should be done after 6 weeks the latest) - can we maybe move things forward also without the 3rd reviewer?
