### 2022-06-29 #4337 @dhhagan
Hi @hechth - I have a few comments/requests before I accept:

  1. It looks like there is a license mismatch. The paper says CC4 and the repo says MIT. Can you make sure those are the same?
  2. Could you add instructions for installing via `pip`?
  3. Am I missing the documentation somewhere? I only see the one notebook file...is there API documentation somewhere?

I will take another look and proceed once I hear back. Let me know if you have any questions!
