### 2022-06-29 28e047e
Added pip installation instructions
