### 2022-06-29 #4337 @hechth
@dhhagan actually the license of the paper is CC4 and the code is MIT.
I added a sentence on how to install the package via pip.
The API documentation is inside the code - it is not hosted somewhere else explicitly at the moment.

If it is required we can setup a readthedocs page.
