### 2022-07-02 #4337 @dhhagan
@hechth Thanks for clearing that up.

Since the reviewers gave the API docs the 'okay', I think it meets the JOSS spec. I do think adding autodocs in the future may be helpful, but this looks good to go for now. To move forward, I will need you to:

  1. Make a final release and archive (make sure the archive deposit has the correct metadata (title and author list))
  2. Report the release version and archive DOI in this thread
