### 2022-07-05 #4337 @hechth
@dhhagan Just to be clear: Make a new release on GitHub, archive the software on Zenodo (I guess) and then state the release version and the DOI link to Zenodo here?
