+++ .github/workflows/publish_pypi.yaml
@@ -13,7 +13,7 @@ jobs:
     - name: Set up Python
       uses: actions/setup-python@v3
       with:
-        python-version: 3.10
+        python-version: 3.7
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
